+++ tests/sparseml/pytorch/sparsification/pruning/test_modifier_pruning_layer.py
@@ -122,44 +122,3 @@ class TestLayerPruningModifier(ScheduledUpdateModifierTest):
             for (name, mod) in modifier._layer_modules.items():
                 assert mod is None
                 assert not isinstance(get_layer(name, model), Identity)
-
-
-def _test_pruning_modifier_serialization_vals(
-    yaml_modifier, serialized_modifier, obj_modifier
-):
-    assert (
-        yaml_modifier.init_sparsity
-        == serialized_modifier.init_sparsity
-        == obj_modifier.init_sparsity
-    )
-    assert (
-        yaml_modifier.final_sparsity
-        == serialized_modifier.final_sparsity
-        == obj_modifier.final_sparsity
-    )
-    assert (
-        yaml_modifier.start_epoch
-        == serialized_modifier.start_epoch
-        == obj_modifier.start_epoch
-    )
-    assert (
-        yaml_modifier.end_epoch
-        == serialized_modifier.end_epoch
-        == obj_modifier.end_epoch
-    )
-    assert (
-        yaml_modifier.update_frequency
-        == serialized_modifier.update_frequency
-        == obj_modifier.update_frequency
-    )
-    assert yaml_modifier.params == serialized_modifier.params == obj_modifier.params
-    assert (
-        yaml_modifier.inter_func
-        == serialized_modifier.inter_func
-        == obj_modifier.inter_func
-    )
-    assert (
-        str(yaml_modifier.mask_type)
-        == str(serialized_modifier.mask_type)
-        == str(obj_modifier.mask_type)
-    )
