+++ tests/sparseml/pytorch/sparsification/pruning/test_modifier_pruning_movement.py
@@ -18,10 +18,7 @@ import pytest
 import torch
 
 from flaky import flaky
-from sparseml.pytorch.sparsification import (
-    StructuredPruningMaskCreator,
-    StructuredPruningModifier,
-)
+from sparseml.pytorch.sparsification import MovementPruningModifier
 from sparseml.pytorch.utils import tensor_sparsity
 from tests.sparseml.pytorch.helpers import LinearNet
 from tests.sparseml.pytorch.optim.test_modifier import (
@@ -31,7 +28,6 @@ from tests.sparseml.pytorch.optim.test_modifier import (
 )
 from tests.sparseml.pytorch.sparsification.pruning.helpers import (
     pruning_modifier_serialization_vals_test,
-    sparsity_mask_creator_test,
 )
 
 
@@ -42,38 +38,6 @@ from tests.sparseml.pytorch.helpers import (  # noqa isort:skip
 )
 
 
-@pytest.mark.skipif(
-    os.getenv("NM_ML_SKIP_PYTORCH_TESTS", False),
-    reason="Skipping pytorch tests",
-)
-@pytest.mark.parametrize(
-    ("tensor_shape,structure_type,tensor_group_idxs"),
-    [
-        ([[64, 64]] * 10, "filter", None),
-        ([[64, 64, 3, 3], [64, 64]], "channel", None),
-        ([[64, 64]] * 6, "filter", [[0, 3], [1, 2, 5], [4]]),
-        ([[64, 64, 3, 3]] * 4, "channel", [[0, 1, 2, 3]]),
-    ],
-)
-@pytest.mark.parametrize("sparsity_val", [0.0, 0.4, 0.6, 0.9, 0.99, 1.0])
-def test_structured_sparsity_mask_creator(
-    tensor_shape, structure_type, tensor_group_idxs, sparsity_val
-):
-    mask_creator = StructuredPruningMaskCreator(
-        structure_type, tensor_group_idxs=tensor_group_idxs
-    )
-    masks = sparsity_mask_creator_test(tensor_shape, mask_creator, sparsity_val, "cpu")
-    target_dim = 1 if structure_type == "filter" else 0
-    for mask in masks:
-        dimension_mask = 1 - torch.all(mask == 0.0, dim=target_dim).float()
-        assert abs(tensor_sparsity(dimension_mask) - sparsity_val) <= 0.01
-    if tensor_group_idxs:
-        for group in tensor_group_idxs:
-            first_mask = masks[group[0]]
-            for mask_idx in group:
-                assert torch.all(masks[mask_idx] == first_mask)
-
-
 @flaky(max_runs=3, min_passes=2)
 @pytest.mark.skipif(
     os.getenv("NM_ML_SKIP_PYTORCH_TESTS", False),
@@ -82,9 +46,8 @@ def test_structured_sparsity_mask_creator(
 @pytest.mark.parametrize(
     "modifier_lambda",
     [
-        lambda: StructuredPruningModifier(
+        lambda: MovementPruningModifier(
             params="__ALL_PRUNABLE__",
-            param_groups=[["seq.fc1.weight", "seq.fc2.weight"]],
             init_sparsity=0.05,
             final_sparsity=0.95,
             start_epoch=10.0,
@@ -101,7 +64,7 @@ def test_structured_sparsity_mask_creator(
     [create_optim_sgd, create_optim_adam],
     scope="function",
 )
-class TestStructuredPruningModifier(ScheduledUpdateModifierTest):
+class TestMovementPruningModifier(ScheduledUpdateModifierTest):
     def test_lifecycle(
         self,
         modifier_lambda,
@@ -113,36 +76,6 @@ class TestStructuredPruningModifier(ScheduledUpdateModifierTest):
         model = model_lambda()
         optimizer = optim_lambda(model)
         self.initialize_helper(modifier, model)
-
-        # test tensor_group_idxs correctness
-        tensor_group_idxs = modifier.mask_creator.tensor_group_idxs
-        param_names_full = [
-            f"{layer_name}.{param_name}"
-            for layer_name, param_name in zip(
-                modifier.module_masks.layer_names, modifier.module_masks.param_names
-            )
-        ]
-        param_names_to_idx = dict(zip(param_names_full, range(len(param_names_full))))
-        expected_tensor_group_idxs = [
-            {param_names_to_idx[param_name] for param_name in param_group}
-            for param_group in modifier.param_groups
-            if len(param_group) > 1
-        ]
-        found_tensor_group_idxs = [
-            set(tensor_group)
-            for tensor_group in tensor_group_idxs
-            if len(tensor_group) > 1
-        ]
-        assert len(expected_tensor_group_idxs) == len(found_tensor_group_idxs)
-        for expected_tensor_group in expected_tensor_group_idxs:
-            assert any(
-                tensor_group == expected_tensor_group
-                for tensor_group in found_tensor_group_idxs
-            )
-
-        if modifier.start_epoch > 0:
-            assert modifier.applied_sparsity is None
-        assert modifier._mask_creator == modifier._module_masks._mask_creator
 
         # check sparsity is not set before
         for epoch in range(int(modifier.start_epoch)):
@@ -214,15 +147,7 @@ class TestStructuredPruningModifier(ScheduledUpdateModifierTest):
             _test_final_sparsity_applied()
 
 
-def test_structured_pruning_yaml():
-    param_groups = [
-        ["param1", "param2"],
-        [
-            "param3",
-            "param4",
-            "param5",
-        ],
-    ]
+def test_movement_pruning_yaml():
     init_sparsity = 0.05
     final_sparsity = 0.8
     start_epoch = 5.0
@@ -230,10 +155,9 @@ def test_structured_pruning_yaml():
     update_frequency = 1.0
     params = "__ALL_PRUNABLE__"
     inter_func = "cubic"
-    mask_type = "filter"
+    mask_type = "block"
     yaml_str = f"""
-    !StructuredPruningModifier
-        param_groups: {param_groups}
+    !MovementPruningModifier
         init_sparsity: {init_sparsity}
         final_sparsity: {final_sparsity}
         start_epoch: {start_epoch}
@@ -243,12 +167,11 @@ def test_structured_pruning_yaml():
         inter_func: {inter_func}
         mask_type: {mask_type}
     """
-    yaml_modifier = StructuredPruningModifier.load_obj(yaml_str)
-    serialized_modifier = StructuredPruningModifier.load_obj(
+    yaml_modifier = MovementPruningModifier.load_obj(yaml_str)
+    serialized_modifier = MovementPruningModifier.load_obj(
         str(yaml_modifier)
-    )  # type: StructuredPruningModifier
-    obj_modifier = StructuredPruningModifier(
-        param_groups=param_groups,
+    )  # type: MovementPruningModifier
+    obj_modifier = MovementPruningModifier(
         init_sparsity=init_sparsity,
         final_sparsity=final_sparsity,
         start_epoch=start_epoch,
@@ -259,12 +182,7 @@ def test_structured_pruning_yaml():
         mask_type=mask_type,
     )
 
-    assert isinstance(yaml_modifier, StructuredPruningModifier)
+    assert isinstance(yaml_modifier, MovementPruningModifier)
     pruning_modifier_serialization_vals_test(
         yaml_modifier, serialized_modifier, obj_modifier
-    )
-    assert (
-        yaml_modifier.param_groups
-        == serialized_modifier.param_groups
-        == obj_modifier.param_groups
     )
